Add custom BootleqSpace webfont for whitespace
This is a font only contains glyph for "space" (U+0020), the purpose is
to make an opinionated appearance: when a paragraph contains "latin",
"space" and "CJK" characters, we want:

- latin in monospace font,
- space doesn't take "1" space, just serves as a small separation,
- CJK goes its own way.

We will use font-family fallback mechanism to achieve this.

The font is hand made with FontForge, and reference some idea/values
from vecel's Geist font.