Enhance link decoration icons in marker tooltip
The paper-clip icon used to have vertical-align problem (positioned a
bit higher than center).

The cause is that inline-block elements were aligned with line box thus
can hardly be just aligned by vertical-align. We still want to make the
whole <a> inlined (instead of using inline-flex) to allow line-wrap at
the middle of the link.

Now turn the added icon into absolute positioned, let it take 100%
height so it can share the same height of the link text.

- Change content to url(...svg) directly, no need to use empty content +
  background-image.

- Use scale transform to adjust the size of icon.
  Here use a rough 0.25lh transform-origin offset, it doesn't make
  perfect sense but works across Firefox and Chrome (where Chrome
  consider the 100% hieght of icon like 200% to the text, if the text
  has been wrapped to 2 lines).