Add temporary alert about incorrect shelter data
Found that shelter "room" and "occupy" data were incorrect due to
upstream provides 可留容犬最大值 as latest value for every year
(including history years).

For example Taipei's max capacity is 840, but it is true ONLY after 2024
march. It should be 450 before that, while www.pet.gov.tw output every
year with 840.

Before we fix it, display this alert box to user.